requirements.txt
PyQt5
PyQt5-stubs
Pillow
python-magic
click
ffpb
